Working checkers. Need to add different KingPiece
The game works, the logic works. The game does not distinguish visuals for kings or regular pieces yet. The king can move the way it should, you just wouldn't know he is a king. I also need to add a finished game and score visuals. The logic is there just the visuals.